feat: extract client components from pages and add title
