+++ ui/app/(campagnes)/campagnes/mission-locale/accompagnement-refuse/page.tsx
@@ -1,34 +1,11 @@
-"use client";
+import { Metadata } from "next";
 
-import { Box, Stack, Typography } from "@mui/material";
+import AccompagnementRefuseClient from "./AccompagnementRefuseClient";
 
-import { _get, _post } from "@/common/httpClient";
-
-import { MissionLocaleQuestion } from "../_components/question";
+export const metadata: Metadata = {
+  title: "Accompagnement refusé | Tableau de bord de l'apprentissage",
+};
 
 export default function Page() {
-  return (
-    <Stack spacing={3}>
-      <Stack justifyContent="end" alignItems="center">
-        <Box
-          component="img"
-          src="/images/support_solid.svg"
-          alt="Accompagner les apprentis"
-          sx={{
-            height: "180px",
-            userSelect: "none",
-            marginBottom: "16px",
-            marginLeft: "auto",
-          }}
-        />
-      </Stack>
-      <Typography variant="h2" sx={{ color: "var(--text-title-blue-france)" }}>
-        Nous avons bien pris en compte votre choix
-      </Typography>
-      <Typography>
-        Nous ne manquerons pas de tenir compte de votre retour. N’hésitez pas à nous recontacter en cas de besoin.
-      </Typography>
-      <MissionLocaleQuestion />
-    </Stack>
-  );
+  return <AccompagnementRefuseClient />;
 }
